pyproject.toml
[project]
name = "pith"
version = "0.1.0"
description = "A minimal, self-extending personal AI agent."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite",
  "httpx",
  "pydantic-ai",
  "prompt-toolkit",
  "PyYAML",
]

[dependency-groups]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[project.scripts]
pith = "pith.cli:main"

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
